test: move some BaseIntegrationTest methods to their correct home.
BaseIntegrationTest methods belong to integration.cc rather than specific _test.

This is useful in #415, since individual _test.cc files become independent test targets and linked objects, there's no need to have http2_integration_test depend on integration_test for example.

This patch is pure code movement.
