+++ test/integration/http2_integration_test.cc
@@ -121,61 +121,6 @@ TEST_F(Http2IntegrationTest, DownstreamResetBeforeResponseComplete) {
   testDownstreamResetBeforeResponseComplete();
 }
 
-void BaseIntegrationTest::testDownstreamResetBeforeResponseComplete() {
-  IntegrationCodecClientPtr codec_client;
-  FakeHttpConnectionPtr fake_upstream_connection;
-  IntegrationStreamDecoderPtr response(new IntegrationStreamDecoder(*dispatcher_));
-  Http::StreamEncoder* downstream_request{};
-  FakeStreamPtr upstream_request;
-  std::list<std::function<void()>> actions = {
-      [&]()
-          -> void { codec_client = makeHttpConnection(HTTP_PORT, Http::CodecClient::Type::HTTP2); },
-      [&]() -> void {
-        downstream_request =
-            &codec_client->startRequest(Http::TestHeaderMapImpl{{":method", "GET"},
-                                                                {":path", "/test/long/url"},
-                                                                {":scheme", "http"},
-                                                                {":authority", "host"},
-                                                                {"cookie", "a=b"},
-                                                                {"cookie", "c=d"}},
-                                        *response);
-        codec_client->sendData(*downstream_request, 0, true);
-      },
-      [&]() -> void {
-        fake_upstream_connection = fake_upstreams_[0]->waitForHttpConnection(*dispatcher_);
-      },
-      [&]() -> void { upstream_request = fake_upstream_connection->waitForNewStream(); },
-      [&]() -> void {
-        upstream_request->waitForEndStream(*dispatcher_);
-        EXPECT_EQ(upstream_request->headers().get(Http::Headers::get().Cookie)->value(),
-                  "a=b; c=d");
-      },
-      [&]() -> void {
-        upstream_request->encodeHeaders(Http::TestHeaderMapImpl{{":status", "200"}}, false);
-        upstream_request->encodeData(512, false);
-      },
-      [&]() -> void { response->waitForBodyData(512); },
-      [&]() -> void { codec_client->sendReset(*downstream_request); }};
-
-  if (fake_upstreams_[0]->httpType() == FakeHttpConnection::Type::HTTP1) {
-    actions.push_back([&]() -> void { fake_upstream_connection->waitForDisconnect(); });
-  } else {
-    actions.push_back([&]() -> void { upstream_request->waitForReset(); });
-    actions.push_back([&]() -> void { fake_upstream_connection->close(); });
-    actions.push_back([&]() -> void { fake_upstream_connection->waitForDisconnect(); });
-  }
-
-  actions.push_back([&]() -> void { codec_client->close(); });
-  executeActions(actions);
-
-  EXPECT_TRUE(upstream_request->complete());
-  EXPECT_EQ(0U, upstream_request->bodyLength());
-
-  EXPECT_FALSE(response->complete());
-  EXPECT_STREQ("200", response->headers().Status()->value().c_str());
-  EXPECT_EQ(512U, response->body().size());
-}
-
 TEST_F(Http2IntegrationTest, BadMagic) {
   Buffer::OwnedImpl buffer("hello");
   std::string response;
@@ -219,58 +164,6 @@ TEST_F(Http2IntegrationTest, GoAway) {
 
   EXPECT_TRUE(response->complete());
   EXPECT_STREQ("200", response->headers().Status()->value().c_str());
-}
-
-void BaseIntegrationTest::testTrailers(uint64_t request_size, uint64_t response_size) {
-  IntegrationCodecClientPtr codec_client;
-  FakeHttpConnectionPtr fake_upstream_connection;
-  IntegrationStreamDecoderPtr response(new IntegrationStreamDecoder(*dispatcher_));
-  Http::StreamEncoder* request_encoder;
-  FakeStreamPtr upstream_request;
-  Http::TestHeaderMapImpl request_trailers{{"request1", "trailer1"}, {"request2", "trailer2"}};
-  Http::TestHeaderMapImpl response_trailers{{"response1", "trailer1"}, {"response2", "trailer2"}};
-  executeActions(
-      {[&]() -> void {
-        codec_client = makeHttpConnection(HTTP_BUFFER_PORT, Http::CodecClient::Type::HTTP2);
-      },
-       [&]() -> void {
-         request_encoder =
-             &codec_client->startRequest(Http::TestHeaderMapImpl{{":method", "POST"},
-                                                                 {":path", "/test/long/url"},
-                                                                 {":scheme", "http"},
-                                                                 {":authority", "host"}},
-                                         *response);
-         codec_client->sendData(*request_encoder, request_size, false);
-         codec_client->sendTrailers(*request_encoder, request_trailers);
-       },
-       [&]() -> void {
-         fake_upstream_connection = fake_upstreams_[0]->waitForHttpConnection(*dispatcher_);
-       },
-       [&]() -> void { upstream_request = fake_upstream_connection->waitForNewStream(); },
-       [&]() -> void { upstream_request->waitForEndStream(*dispatcher_); },
-       [&]() -> void {
-         upstream_request->encodeHeaders(Http::TestHeaderMapImpl{{":status", "200"}}, false);
-         upstream_request->encodeData(response_size, false);
-         upstream_request->encodeTrailers(response_trailers);
-       },
-       [&]() -> void { response->waitForEndStream(); },
-       // Cleanup both downstream and upstream
-       [&]() -> void { codec_client->close(); },
-       [&]() -> void { fake_upstream_connection->close(); },
-       [&]() -> void { fake_upstream_connection->waitForDisconnect(); }});
-
-  EXPECT_TRUE(upstream_request->complete());
-  EXPECT_EQ(request_size, upstream_request->bodyLength());
-  if (fake_upstreams_[0]->httpType() == FakeHttpConnection::Type::HTTP2) {
-    EXPECT_THAT(*upstream_request->trailers(), HeaderMapEqualRef(&request_trailers));
-  }
-
-  EXPECT_TRUE(response->complete());
-  EXPECT_STREQ("200", response->headers().Status()->value().c_str());
-  EXPECT_EQ(response_size, response->body().size());
-  if (fake_upstreams_[0]->httpType() == FakeHttpConnection::Type::HTTP2) {
-    EXPECT_THAT(*response->trailers(), HeaderMapEqualRef(&response_trailers));
-  }
 }
 
 TEST_F(Http2IntegrationTest, Trailers) { testTrailers(1024, 2048); }
